pyproject.toml
[project]
name = "pln-qa"
version = "0.1.0"
description = "Projeto QA para a disciplina de Processamento de Linguagem Natural"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub[hf-transfer]>=0.34.4",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
    "transformers>=4.55.0",
]
